+++ packages/terra-framework-docs/src/terra-dev-site/test/data-grid/flowsheet-data-grid/RowSelection.test.jsx
@@ -99,7 +99,10 @@ const RowSelection = () => {
     }
   }, [clearSelectedRow, rowData, selectedRow]);
 
-  const onCellSelect = useCallback((rowId, columnId, isMetaPressed) => {
+  const onCellSelect = useCallback((selectedCell) => {
+    const { rowId } = selectedCell;
+    const { columnId } = selectedCell;
+    const { isMetaPressed } = selectedCell;
     if (rowId && columnId) {
       const rowIndex = rowData.findIndex(e => e.id === rowId);
       const columnIndex = cols.findIndex(e => e.id === columnId);