+++ packages/terra-table/src/subcomponents/ColumnHeader.jsx
@@ -132,7 +132,6 @@ const ColumnHeader = (props) => {
 
   // Is needed to adjust the header column resize handler to accommodate actions header height
   const initialHeight = hasColumnHeaderActions ? `${headerContainerHeight}px` : undefined;
-
   return (
     <thead ref={headerRef}>
       <tr
@@ -141,6 +140,7 @@ const ColumnHeader = (props) => {
         className={cx('column-header-row', { hidden: !hasVisibleColumnHeaders })}
         height={hasVisibleColumnHeaders ? headerHeight : undefined}
       >
+
         {columns.map((column, columnIndex) => (
           <ColumnHeaderCell
             key={`${column.id}-headerCell`}
@@ -148,6 +148,7 @@ const ColumnHeader = (props) => {
             tableId={tableId}
             columnId={column.id}
             columnIndex={columnIndex}
+            columnSpan={column.columnSpan}
             displayName={column.displayName}
             isDisplayVisible={column.isDisplayVisible}
             width={column.width}
@@ -190,6 +191,7 @@ const ColumnHeader = (props) => {
               isActionCell
               action={column.action}
               columnIndex={columnIndex}
+              columnSpan={column.columnSpan}
               isDisplayVisible={column.isDisplayVisible}
               width={column.isResizable && column.width}
               minimumWidth={column.minimumWidth}